Add pickup distance to PlayerInteractionController
Introduced a new 'pickupDistance' variable in PlayerInteractionController to provide a limit to how far a player can interact with objects. This aims to add a realistic constraint to gameplay, where players need to be within a certain range before they can interact with objects. This change has been reflected in the 'UpdateTarget' method, where a maxDistance parameter has been added to the raycast operation to stop highlighting of interactables beyond the pickup distance.