Add bobbing effect to held objects
Updated player interaction mechanics to include a bobbing effect when held objects are moved. This is achieved by altering the y position with a sine function using the field BobbingIntensity. This makes the interaction more noticeable and attractive for a player. Adjusted outline width and color in the WeirdOutlineMaterial.mat for better visualization. Made necessary changes in Hold() and Release() methods in InteractableObjectcs. Adjusted related parameters and method calls in GameScene.unity and PlayerInteractionController.cs.