Add shipping area and box animation
A new "ShippingArea" component was created as well as a new animation for shipped boxes in the "InteractableObject" component. Now, when a box collides with the ShippingArea, it plays a shipped animation and fires a BoxDelivered message. These changes improve the game by visualizing a box's departure from gameplay and alerting any applicable handlers in the system. Also, support utilities and scripts were added to allow game objects to determine if they have a specific component.