Add ShippingArea prefab and update game logic
Created a new prefab for the ShippingArea, transitioning individual game objects in the GameScene to use the prefab instead. This change simplifies the creation and maintenance of ShippingArea instances. Also, added a new KPI "BoxShipped" to the GameState and updated ShippingArea logic to increment this KPI when a box is delivered. Aligned coworker data in GameState to use Ids instead names for better data management. Updated render queue on 'MAT_Boxes_Large' material to ensure proper rendering order

Closes #92.