Add throw object functionality and effects on player interaction
Introduced throwing functionality to the player's interaction with objects in the game. A slider was added for visual representation of throw force progression when throw action is initiated by the player. Various game object aspects, such as the position and size of objects being thrown, along with some underlying scripts were updated. An additional mechanism for adjustable throwing force was also introduced, leading to enhanced gameplay and a more immersive experience provided to the player.