Add new materials, props items and a scene
This commit introduces new game assets. A 'blue' material was created and later used in the modification of a 'BoxTeleporter' prefab which was added into the 'Props Items'. In addition, a new scene called 'Warehouse' was initiated under 'Chaos'.

These additions are a part of the process of building a diverse collection of game assets and scenes, facilitating varied experiences for the player and ensuring that developers have a broad range of elements for designing game levels.