Refine game objects interactions and properties
Several changes were made to improve the game's mechanics and interactions. Firstly, unnecessary prefab instances were removed from the PlayerPlayground scene for better efficiency. Furthermore, BoxDelivered logic was extended in the BoxDelivered.cs script to include the box color, refining the way boxes are handled and delivered. In the ShippingArea.cs script, color validation was added to ensure only boxes with the correct color are processed. Lastly, the physics of the JumpPad was improved by setting the y-component of the player's velocity to zero before the jump force is applied for a more uniform jump behavior.